Revert package moving for secret related files
