+++ pkg/ddc/alluxio/ufs_internal.go
@@ -17,8 +17,6 @@ package alluxio
 
 import (
 	"fmt"
-	"github.com/fluid-cloudnative/fluid/pkg/utils/kubeclient"
-
 	datav1alpha1 "github.com/fluid-cloudnative/fluid/api/v1alpha1"
 	"github.com/fluid-cloudnative/fluid/pkg/common"
 	"github.com/fluid-cloudnative/fluid/pkg/ddc/alluxio/operations"
@@ -158,7 +156,7 @@ func (e *AlluxioEngine) genUFSMountOptions(m datav1alpha1.Mount) (map[string]str
 	for _, item := range m.EncryptOptions {
 
 		sRef := item.ValueFrom.SecretKeyRef
-		secret, err := kubeclient.GetSecret(e.Client, sRef.Name, e.namespace)
+		secret, err := utils.GetSecret(e.Client, sRef.Name, e.namespace)
 		if err != nil {
 			e.Log.Error(err, "get secret by mount encrypt options failed", "name", item.Name)
 			return mOptions, err
